Move synchronization logic to sync.service.ts
