+++ src/app/components/photoboothImport/photoboothImport.component.html
@@ -115,6 +115,6 @@
 
   </table>
 
-  <button mat-raised-button (click)="getLocalProjects()">SELECT PROJECTS</button>
+  <button mat-raised-button (click)="getSyncProjects()">SELECT PROJECTS</button>
 
 </div>
