+++ src/app/services/sync/sync.service.ts
@@ -1,7 +1,7 @@
 import {Injectable} from '@angular/core';
 import {AngularFirestore, AngularFirestoreCollection} from "@angular/fire/firestore";
-import {IProject} from "../../shared/interfaces/project";
-import {Subject} from "rxjs";
+import {IProject, ISyncProject} from "../../shared/interfaces/project";
+import {Observable, Subject, Subscription} from "rxjs";
 import firebase from "firebase";
 import Timestamp = firebase.firestore.Timestamp;
 import {ElectronService} from "../../core/services";
@@ -9,45 +9,66 @@ import {DirectoryTreeService} from "../directory-tree/directory-tree.service";
 import {AngularFireStorage} from "@angular/fire/storage";
 import TaskState = firebase.storage.TaskState;
 import {IDirectoryTree} from "../../shared/interfaces/directory-tree";
-import {IFile} from "../../shared/interfaces/file";
+import {IFile, ISyncFile} from "../../shared/interfaces/file";
+import {DialogData} from "../../shared/interfaces/dialog-data";
+import {MatDialog} from "@angular/material/dialog";
+import {SyncDialogComponent} from "../../components/sync-dialog/sync-dialog.component";
+import {MatSnackBar} from "@angular/material/snack-bar";
 
 
 @Injectable({
   providedIn: 'root'
 })
 export class SyncService {
+  subscriptions: Array<Subscription>;
   projectsCollection: AngularFirestoreCollection<IProject>;
+  projectsDirectory: string;
+  remoteProjects: Array<IProject>;
   remoteProjects$: Subject<Array<IProject>>;
+  localProjects: Array<IProject>;
   localProjects$: Subject<Array<IProject>>;
+  syncProjects$: Subject<Array<ISyncProject>>;
 
   constructor(private afs: AngularFirestore,
               private storage: AngularFireStorage,
               private electronService: ElectronService,
-              private directoryTreeService: DirectoryTreeService) {
+              private directoryTreeService: DirectoryTreeService,
+              private dialog: MatDialog,
+              private snackBar: MatSnackBar) {
+    this.initialize();
+  }
+
+  initialize(): void {
+    this.subscriptions = new Array<Subscription>();
     this.projectsCollection = this.afs.collection<IProject>('projects');
+    this.projectsDirectory = "";
+    this.remoteProjects = new Array<IProject>();
     this.remoteProjects$ = new Subject<Array<IProject>>();
+    this.localProjects = new Array<IProject>();
     this.localProjects$ = new Subject<Array<IProject>>();
+    this.syncProjects$ = new Subject<Array<ISyncProject>>();
+
+    this.subscriptions.push(
+      this.localProjects$.subscribe((projects) => {
+        this.localProjects = projects;
+        this.getRemoteProjects();
+      })
+    );
+
+    this.subscriptions.push(
+      this.remoteProjects$.subscribe((projects) => {
+        this.remoteProjects = projects;
+        this.compareProjects();
+      })
+    );
   }
 
-  getRemoteProjects(): void {
-    const projects = new Array<IProject>();
-    const projectsSubscription = this.projectsCollection.get()
-      .subscribe(
-        (querySnapshot) => {
-          querySnapshot.forEach(doc => projects.push(doc.data()));
-          this.remoteProjects$.next(projects);
-        },
-        (error) => {
-          console.log(error);
-        },
-        () => {
-          console.log('The list of remote projects downloaded successfully.');
-          projectsSubscription.unsubscribe();
-        }
-      );
+  async getSyncProjects(directoryPath: string): Promise<void> {
+    this.projectsDirectory = directoryPath;
+    await this.getLocalProjects(directoryPath);
   }
 
-  async getLocalProjects(directoryPath: string): Promise<void> {
+  private async getLocalProjects(directoryPath: string): Promise<void> {
     const projects = new Array<IProject>();
 
     const directoryChildren = this.electronService.fs.readdirSync(directoryPath);
@@ -81,7 +102,278 @@ export class SyncService {
     this.localProjects$.next(projects);
   }
 
-  uploadFile(projectsDirectory: string, file: IFile, isConflictingFile?: boolean): Promise<void> {
+  private getRemoteProjects(): void {
+    const projects = new Array<IProject>();
+    const projectsSubscription = this.projectsCollection.get()
+      .subscribe(
+        (querySnapshot) => {
+          querySnapshot.forEach(doc => projects.push(doc.data()));
+          this.remoteProjects$.next(projects);
+        },
+        (error) => {
+          console.log(error);
+        },
+        () => {
+          console.log('The list of remote projects downloaded successfully.');
+          projectsSubscription.unsubscribe();
+        }
+      );
+  }
+
+  private compareProjects(): void {
+    const projects = new Array<ISyncProject>();
+
+    this.remoteProjects.forEach((remoteProject) => {
+      const localProjectIndex = this.localProjects.findIndex(localProject => localProject.name == remoteProject.name);
+
+      if (localProjectIndex == -1) {
+        projects.push(
+          {
+            name: remoteProject.name,
+            creationDate: remoteProject.creationDate,
+            directoryTree: remoteProject.directoryTree,
+            files: remoteProject.files.map(file => {
+              return {
+                ...file,
+                sync: 'cloud'
+              };
+            }),
+            sync: 'cloud'
+          }
+        );
+      } else {
+        const files = new Array<ISyncFile>();
+
+        remoteProject.files.forEach((remoteFile) => {
+          const localFileIndex = this.localProjects[localProjectIndex].files
+            .findIndex(localFile => localFile.name == remoteFile.name && localFile.path == remoteFile.path);
+
+          if (localFileIndex == -1) {
+            files.push(
+              {
+                ...remoteFile,
+                sync: 'cloud'
+              }
+            );
+          } else {
+            if (remoteFile.sha256 == this.localProjects[localProjectIndex].files[localFileIndex].sha256) {
+              files.push(
+                {
+                  ...remoteFile,
+                  sync: 'synchronized'
+                }
+              );
+            } else {
+              files.push(
+                {
+                  ...remoteFile,
+                  sync: 'unsynchronized'
+                }
+              );
+            }
+          }
+        });
+
+        projects.push({
+          ...remoteProject,
+          files: files,
+          sync: null
+        });
+      }
+    });
+
+    this.localProjects.forEach(localProject => {
+      const remoteProjectIndex = this.remoteProjects.findIndex(remoteProject => remoteProject.name == localProject.name);
+
+      if (remoteProjectIndex == -1) {
+        projects.push(
+          {
+            name: localProject.name,
+            creationDate: localProject.creationDate,
+            directoryTree: localProject.directoryTree,
+            files: localProject.files.map(file => {
+              return {
+                ...file,
+                sync: 'local'
+              };
+            }),
+            sync: 'local'
+          }
+        );
+      } else {
+        localProject.files.forEach((localFile) => {
+          const remoteFileIndex = this.remoteProjects[remoteProjectIndex].files
+            .findIndex(remoteFile => remoteFile.name == localFile.name && remoteFile.path == localFile.path);
+
+          if (remoteFileIndex == -1) {
+            projects.find(syncProject => syncProject.name == localProject.name).files
+              .push({
+                ...localFile,
+                sync: 'local'
+              });
+          }
+        });
+      }
+    });
+
+    projects.forEach(syncProject => {
+      if (syncProject.sync == null) {
+        SyncService.updateProjectSynchronizationState(syncProject);
+      }
+    });
+
+    this.syncProjects$.next(projects);
+  }
+
+  syncProject(project: ISyncProject, event: MouseEvent): void {
+    // Prevent project expansion / collapse when we click on a sync icon.
+    event.stopPropagation();
+
+    const dialogData: DialogData = {
+      title: "Project synchronization",
+      content: `You are about to synchronize the entire "${project.name}" project. Do you want to continue?`,
+      action: "Synchronize"
+    };
+
+    this.openDialog(dialogData).toPromise()
+      .then(async userAction => {
+        if (userAction) {
+          try {
+            await this.downloadFiles(project, project.files.filter(file => file.sync == 'cloud'));
+            await this.uploadFiles(project, project.files.filter(file => file.sync == 'local'));
+            if (project.sync == 'unsynchronized') {
+              this.showSnackBar('Some synchronization conflicts have been detected, you must resolve them manually.');
+            }
+          } catch (error) {
+            console.log(error);
+          }
+        }
+      })
+      .catch(error => console.log(error));
+  }
+
+  downloadProject(project: ISyncProject, event: MouseEvent): void {
+    // Prevent project expansion / collapse when we click on a sync icon.
+    event.stopPropagation();
+
+    const dialogData: DialogData = {
+      title: "Project download",
+      content: `You are about to download the entire "${project.name}" project. Do you want to continue?`,
+      action: "Download"
+    };
+
+    this.openDialog(dialogData).toPromise()
+      .then(async userAction => {
+        if (userAction) {
+          try {
+            await this.downloadFiles(project, project.files);
+          } catch (error) {
+            console.log(error);
+          }
+        }
+      })
+      .catch(error => console.log(error));
+  }
+
+  uploadProject(project: ISyncProject, event: MouseEvent): void {
+    // Prevent project expansion / collapse when we click on a sync icon.
+    event.stopPropagation();
+
+    const dialogData: DialogData = {
+      title: "Project upload",
+      content: `You are about to upload the entire "${project.name}" project. Do you want to continue?`,
+      action: "Upload"
+    };
+
+    this.openDialog(dialogData).toPromise()
+      .then(async userAction => {
+        if (userAction) {
+          try {
+            await this.uploadFiles(project, project.files);
+          } catch (error) {
+            console.log(error);
+          }
+        }
+      })
+      .catch(error => console.log(error));
+  }
+
+  syncFile(project: ISyncProject, file: ISyncFile): void {
+    const remoteFile = this.remoteProjects
+      .find(remoteProject => remoteProject.name == project.name).files
+      .find(remoteFile => remoteFile.name == file.name);
+
+    const localFile = this.localProjects
+      .find(localProject => localProject.name == project.name).files
+      .find(localFile => localFile.name == file.name);
+
+    const conflictingFiles = [
+      `Server version: ${remoteFile.name} ; created: ${remoteFile.creationDate.toDate().toLocaleDateString()} at ${remoteFile.creationDate.toDate().toLocaleTimeString()}`,
+      `Local version: ${localFile.name} ; created: ${localFile.creationDate.toDate().toLocaleDateString()} at ${localFile.creationDate.toDate().toLocaleTimeString()}`
+    ];
+
+    const dialogData: DialogData = {
+      title: "File synchronization",
+      content: "A synchronization conflict is detected. You must resolve it manually by choosing the version of the file to keep.",
+      action: "Synchronize",
+      conflictingFiles: conflictingFiles
+    };
+
+    this.openDialog(dialogData).toPromise()
+      .then(async userAction => {
+        if (userAction) {
+          if (conflictingFiles.indexOf(userAction as string) == 0) {
+            try {
+              file.creationDate = remoteFile.creationDate;
+              file.size = remoteFile.size;
+              file.sha256 = remoteFile.sha256;
+              await this.downloadFiles(project, [file], true);
+            } catch (error) {
+              console.log(error);
+            }
+          } else {
+            try {
+              file.creationDate = localFile.creationDate;
+              file.size = localFile.size;
+              file.sha256 = localFile.sha256;
+              await this.uploadFiles(project, [file], true);
+            } catch (error) {
+              console.log(error);
+            }
+          }
+        }
+      })
+      .catch(error => console.log(error));
+  }
+
+  async uploadFiles(project: ISyncProject, files: Array<ISyncFile>, areConflictingFiles?: boolean): Promise<void> {
+    project.sync = 'isSynchronizing';
+    files.forEach(file => file.sync = 'isSynchronizing');
+
+    for (const file of files) {
+      try {
+        const fileToUpload: IFile = {
+          name: file.name,
+          creationDate: file.creationDate,
+          path: file.path,
+          size: file.size,
+          sha256: file.sha256
+        };
+
+        await this.uploadFile(this.projectsDirectory, fileToUpload, areConflictingFiles);
+        file.sync = 'synchronized';
+      } catch (error) {
+        areConflictingFiles
+          ? file.sync = 'unsynchronized'
+          : file.sync = 'local';
+        this.showSnackBar(`Error: the upload of the file "${file.name}" failed!`);
+        console.log(error);
+      }
+    }
+    SyncService.updateProjectSynchronizationState(project);
+  }
+
+  private uploadFile(projectsDirectory: string, file: IFile, isConflictingFile?: boolean): Promise<void> {
     const fullPath = this.electronService.path.join(projectsDirectory, file.path);
 
     return new Promise<void>((resolve, reject) => {
@@ -108,7 +400,7 @@ export class SyncService {
     });
   }
 
-  updateProjectStructure(projectsDirectory: string, file: IFile, isConflictingFile?: boolean): Promise<void> {
+  private updateProjectStructure(projectsDirectory: string, file: IFile, isConflictingFile?: boolean): Promise<void> {
     const pathSegments = file.path.split(this.electronService.path.sep);
     const projectName = pathSegments[0];
     const projectPath = this.electronService.path.join(projectsDirectory, projectName);
@@ -164,7 +456,7 @@ export class SyncService {
     });
   }
 
-  updateDirectoryTree(directoryTree: IDirectoryTree, pathSegments: Array<string>, filePath: string): void {
+  private updateDirectoryTree(directoryTree: IDirectoryTree, pathSegments: Array<string>, filePath: string): void {
     pathSegments = pathSegments.slice(1);
 
     if (pathSegments.length == 1) {
@@ -186,7 +478,26 @@ export class SyncService {
     }
   }
 
-  downloadFile(projectsDirectory: string, file: IFile): Promise<void> {
+  async downloadFiles(project: ISyncProject, files: Array<ISyncFile>, areConflictingFiles?: boolean): Promise<void> {
+    project.sync = 'isSynchronizing';
+    files.forEach(file => file.sync = 'isSynchronizing');
+
+    for (const file of files) {
+      try {
+        await this.downloadFile(this.projectsDirectory, file);
+        file.sync = 'synchronized';
+      } catch (error) {
+        areConflictingFiles
+          ? file.sync = 'unsynchronized'
+          : file.sync = 'cloud';
+        this.showSnackBar(`Error: the download of the file "${file.name}" failed!`);
+        console.log(error);
+      }
+    }
+    SyncService.updateProjectSynchronizationState(project);
+  }
+
+  private downloadFile(projectsDirectory: string, file: IFile): Promise<void> {
     return new Promise<void>((resolve, reject) => {
       this.storage.ref(file.path).getDownloadURL().toPromise()
         .then((url) => {
@@ -220,5 +531,35 @@ export class SyncService {
         })
         .catch(error => reject(error));
     });
+  }
+
+  private static updateProjectSynchronizationState(syncProject: ISyncProject): void {
+    if (syncProject.files.every(file => file.sync == 'synchronized')) {
+      syncProject.sync = 'synchronized';
+    } else if (syncProject.files.every(file => file.sync == 'cloud')) {
+      syncProject.sync = 'cloud';
+    } else if (syncProject.files.every(file => file.sync == 'local')) {
+      syncProject.sync = 'local';
+    } else {
+      syncProject.sync = 'unsynchronized';
+    }
+  }
+
+  private openDialog(data: DialogData): Observable<any> {
+    const dialogRef = this.dialog.open(SyncDialogComponent, {data: data});
+    return dialogRef.afterClosed();
+  }
+
+  showSnackBar(message: string, event?: MouseEvent): void {
+    // Prevent project expansion / collapse when we click on a sync icon.
+    if (event) {
+      event.stopPropagation();
+    }
+
+    this.snackBar.open(message, null, {duration: 2000});
+  }
+
+  ngOnDestroy(): void {
+    this.subscriptions.forEach(subscription => subscription.unsubscribe());
   }
 }
